Prevents fatal exception on calling DisconnectAsync
This line:

_backingDatagramSocket.MessageReceived -= DatagramMessageReceived;

causes this exception (not throwing in NuGet package at all):

A method was called at an unexpected time.

and will result in very unexpected behaviors in the next operations
(even other TCP, UDP calls).
Wrapping it in a try catch will resolve this problem by handling the
exception. The root cause still to be determined.